@@ -1,14 +1,24 @@
 module github.com/percybolmer/grpcstreams
 
-go 1.15
+go 1.20
 
 require (
-	github.com/golang/protobuf v1.4.3
-	github.com/improbable-eng/grpc-web v0.13.0
-	github.com/mackerelio/go-osstat v0.1.0
-	github.com/mwitkow/go-conntrack v0.0.0-20190716064945-2f068394615f // indirect
-	github.com/percybolmer/grpcexample v0.0.0-20201226070717-225166c4cbe6 // indirect
-	google.golang.org/grpc v1.35.0
-	google.golang.org/protobuf v1.25.0
-	grpc.go4.org v0.0.0-20170609214715-11d0a25b4919
+	github.com/improbable-eng/grpc-web v0.15.0
+	github.com/mackerelio/go-osstat v0.2.5
+	google.golang.org/grpc v1.64.0
+	google.golang.org/protobuf v1.34.2
+)
+
+require (
+	github.com/cenkalti/backoff/v4 v4.1.1 // indirect
+	github.com/desertbit/timer v0.0.0-20180107155436-c41aec40b27f // indirect
+	github.com/golang/protobuf v1.5.4 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/klauspost/compress v1.11.7 // indirect
+	github.com/rs/cors v1.7.0 // indirect
+	golang.org/x/net v0.22.0 // indirect
+	golang.org/x/sys v0.20.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	google.golang.org/genproto v0.0.0-20210126160654-44e461bb6506 // indirect
+	nhooyr.io/websocket v1.8.6 // indirect
 )
