@@ -1,13 +1,12 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.25.0
-// 	protoc        v3.13.0
+// 	protoc-gen-go v1.34.2
+// 	protoc        v3.4.0
 // source: service.proto
 
 package hardwaremonitoring
 
 import (
-	proto "github.com/golang/protobuf/proto"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -20,10 +19,6 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
-
-// This is a compile-time assertion that a sufficiently up-to-date version
-// of the legacy proto package is being used.
-const _ = proto.ProtoPackageIsVersion4
 
 // HardwareStats is a struct containing information about the systems Memory usage and CPU
 type HardwareStats struct {
@@ -162,7 +157,7 @@ func file_service_proto_rawDescGZIP() []byte {
 }
 
 var file_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_service_proto_goTypes = []interface{}{
+var file_service_proto_goTypes = []any{
 	(*HardwareStats)(nil), // 0: main.HardwareStats
 	(*EmptyRequest)(nil),  // 1: main.EmptyRequest
 }
@@ -182,7 +177,7 @@ func file_service_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*HardwareStats); i {
 			case 0:
 				return &v.state
@@ -194,7 +189,7 @@ func file_service_proto_init() {
 				return nil
 			}
 		}
-		file_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*EmptyRequest); i {
 			case 0:
 				return &v.state
