@@ -1,4 +1,8 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
+// versions:
+// - protoc-gen-go-grpc v1.4.0
+// - protoc             v3.4.0
+// source: service.proto
 
 package hardwaremonitoring
 
@@ -11,11 +15,18 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.62.0 or later.
+const _ = grpc.SupportPackageIsVersion8
+
+const (
+	HardwareMonitor_Monitor_FullMethodName = "/main.HardwareMonitor/Monitor"
+)
 
 // HardwareMonitorClient is the client API for HardwareMonitor service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// HardwareMonitor is our service, the service is the API, and rpc statements below will be methods that are available by the API
 type HardwareMonitorClient interface {
 	// Monitor will output stats about the hardware on the system host
 	Monitor(ctx context.Context, in *EmptyRequest, opts ...grpc.CallOption) (HardwareMonitor_MonitorClient, error)
@@ -30,11 +41,12 @@ func NewHardwareMonitorClient(cc grpc.ClientConnInterface) HardwareMonitorClient
 }
 
 func (c *hardwareMonitorClient) Monitor(ctx context.Context, in *EmptyRequest, opts ...grpc.CallOption) (HardwareMonitor_MonitorClient, error) {
-	stream, err := c.cc.NewStream(ctx, &_HardwareMonitor_serviceDesc.Streams[0], "/main.HardwareMonitor/Monitor", opts...)
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &HardwareMonitor_ServiceDesc.Streams[0], HardwareMonitor_Monitor_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &hardwareMonitorMonitorClient{stream}
+	x := &hardwareMonitorMonitorClient{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -64,6 +76,8 @@ func (x *hardwareMonitorMonitorClient) Recv() (*HardwareStats, error) {
 // HardwareMonitorServer is the server API for HardwareMonitor service.
 // All implementations must embed UnimplementedHardwareMonitorServer
 // for forward compatibility
+//
+// HardwareMonitor is our service, the service is the API, and rpc statements below will be methods that are available by the API
 type HardwareMonitorServer interface {
 	// Monitor will output stats about the hardware on the system host
 	Monitor(*EmptyRequest, HardwareMonitor_MonitorServer) error
@@ -87,7 +101,7 @@ type UnsafeHardwareMonitorServer interface {
 }
 
 func RegisterHardwareMonitorServer(s grpc.ServiceRegistrar, srv HardwareMonitorServer) {
-	s.RegisterService(&_HardwareMonitor_serviceDesc, srv)
+	s.RegisterService(&HardwareMonitor_ServiceDesc, srv)
 }
 
 func _HardwareMonitor_Monitor_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -95,7 +109,7 @@ func _HardwareMonitor_Monitor_Handler(srv interface{}, stream grpc.ServerStream)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(HardwareMonitorServer).Monitor(m, &hardwareMonitorMonitorServer{stream})
+	return srv.(HardwareMonitorServer).Monitor(m, &hardwareMonitorMonitorServer{ServerStream: stream})
 }
 
 type HardwareMonitor_MonitorServer interface {
@@ -111,7 +125,10 @@ func (x *hardwareMonitorMonitorServer) Send(m *HardwareStats) error {
 	return x.ServerStream.SendMsg(m)
 }
 
-var _HardwareMonitor_serviceDesc = grpc.ServiceDesc{
+// HardwareMonitor_ServiceDesc is the grpc.ServiceDesc for HardwareMonitor service.
+// It's only intended for direct use with grpc.RegisterService,
+// and not to be introspected or modified (even as a copy)
+var HardwareMonitor_ServiceDesc = grpc.ServiceDesc{
 	ServiceName: "main.HardwareMonitor",
 	HandlerType: (*HardwareMonitorServer)(nil),
 	Methods:     []grpc.MethodDesc{},
