@@ -4,7 +4,11 @@
  * @public
  */
 
-// GENERATED CODE -- DO NOT EDIT!
+// Code generated by protoc-gen-grpc-web. DO NOT EDIT.
+// versions:
+// 	protoc-gen-grpc-web v1.5.0
+// 	protoc              v3.4.0
+// source: service.proto
 
 
 /* eslint-disable */
@@ -21,7 +25,7 @@ proto.main = require('./service_pb.js');
 /**
  * @param {string} hostname
  * @param {?Object} credentials
- * @param {?Object} options
+ * @param {?grpc.web.ClientOptions} options
  * @constructor
  * @struct
  * @final
@@ -29,7 +33,7 @@ proto.main = require('./service_pb.js');
 proto.main.HardwareMonitorClient =
     function(hostname, credentials, options) {
   if (!options) options = {};
-  options['format'] = 'text';
+  options.format = 'text';
 
   /**
    * @private @const {!grpc.web.GrpcWebClientBase} The client
@@ -39,7 +43,7 @@ proto.main.HardwareMonitorClient =
   /**
    * @private @const {string} The hostname
    */
-  this.hostname_ = hostname;
+  this.hostname_ = hostname.replace(/\/+$/, '');
 
 };
 
@@ -47,7 +51,7 @@ proto.main.HardwareMonitorClient =
 /**
  * @param {string} hostname
  * @param {?Object} credentials
- * @param {?Object} options
+ * @param {?grpc.web.ClientOptions} options
  * @constructor
  * @struct
  * @final
@@ -55,7 +59,7 @@ proto.main.HardwareMonitorClient =
 proto.main.HardwareMonitorPromiseClient =
     function(hostname, credentials, options) {
   if (!options) options = {};
-  options['format'] = 'text';
+  options.format = 'text';
 
   /**
    * @private @const {!grpc.web.GrpcWebClientBase} The client
@@ -65,7 +69,7 @@ proto.main.HardwareMonitorPromiseClient =
   /**
    * @private @const {string} The hostname
    */
-  this.hostname_ = hostname;
+  this.hostname_ = hostname.replace(/\/+$/, '');
 
 };
 
@@ -93,27 +97,8 @@ const methodDescriptor_HardwareMonitor_Monitor = new grpc.web.MethodDescriptor(
 
 
 /**
- * @const
- * @type {!grpc.web.AbstractClientBase.MethodInfo<
- *   !proto.main.EmptyRequest,
- *   !proto.main.HardwareStats>}
- */
-const methodInfo_HardwareMonitor_Monitor = new grpc.web.AbstractClientBase.MethodInfo(
-  proto.main.HardwareStats,
-  /**
-   * @param {!proto.main.EmptyRequest} request
-   * @return {!Uint8Array}
-   */
-  function(request) {
-    return request.serializeBinary();
-  },
-  proto.main.HardwareStats.deserializeBinary
-);
-
-
-/**
  * @param {!proto.main.EmptyRequest} request The request proto
- * @param {?Object<string, string>} metadata User defined
+ * @param {?Object<string, string>=} metadata User defined
  *     call metadata
  * @return {!grpc.web.ClientReadableStream<!proto.main.HardwareStats>}
  *     The XHR Node Readable Stream
@@ -130,7 +115,7 @@ proto.main.HardwareMonitorClient.prototype.monitor =
 
 /**
  * @param {!proto.main.EmptyRequest} request The request proto
- * @param {?Object<string, string>} metadata User defined
+ * @param {?Object<string, string>=} metadata User defined
  *     call metadata
  * @return {!grpc.web.ClientReadableStream<!proto.main.HardwareStats>}
  *     The XHR Node Readable Stream
