@@ -1,4 +1,3 @@
-// source: service.proto
 /**
  * @fileoverview
  * @enhanceable
@@ -14,6 +13,7 @@ var global = Function('return this')();
 
 goog.exportSymbol('proto.main.EmptyRequest', null, global);
 goog.exportSymbol('proto.main.HardwareStats', null, global);
+
 /**
  * Generated by JsPbCodeGenerator.
  * @param {Array=} opt_data Optional initial data array, typically from a
@@ -29,47 +29,19 @@ proto.main.HardwareStats = function(opt_data) {
 };
 goog.inherits(proto.main.HardwareStats, jspb.Message);
 if (goog.DEBUG && !COMPILED) {
-  /**
-   * @public
-   * @override
-   */
   proto.main.HardwareStats.displayName = 'proto.main.HardwareStats';
 }
-/**
- * Generated by JsPbCodeGenerator.
- * @param {Array=} opt_data Optional initial data array, typically from a
- * server response, or constructed directly in Javascript. The array is used
- * in place and becomes part of the constructed object. It is not cloned.
- * If no data is provided, the constructed object will be empty, but still
- * valid.
- * @extends {jspb.Message}
- * @constructor
- */
-proto.main.EmptyRequest = function(opt_data) {
-  jspb.Message.initialize(this, opt_data, 0, -1, null, null);
-};
-goog.inherits(proto.main.EmptyRequest, jspb.Message);
-if (goog.DEBUG && !COMPILED) {
-  /**
-   * @public
-   * @override
-   */
-  proto.main.EmptyRequest.displayName = 'proto.main.EmptyRequest';
-}
-
 
 
 if (jspb.Message.GENERATE_TO_OBJECT) {
 /**
- * Creates an object representation of this proto.
+ * Creates an object representation of this proto suitable for use in Soy templates.
  * Field names that are reserved in JavaScript and will be renamed to pb_name.
- * Optional fields that are not set will be set to undefined.
  * To access a reserved field use, foo.pb_<name>, eg, foo.pb_default.
  * For the list of reserved names please see:
- *     net/proto2/compiler/js/internal/generator.cc#kKeyword.
- * @param {boolean=} opt_includeInstance Deprecated. whether to include the
- *     JSPB instance for transitional soy proto support:
- *     http://goto/soy-param-migration
+ *     com.google.apps.jspb.JsClassTemplate.JS_RESERVED_WORDS.
+ * @param {boolean=} opt_includeInstance Whether to include the JSPB instance
+ *     for transitional soy proto support: http://goto/soy-param-migration
  * @return {!Object}
  */
 proto.main.HardwareStats.prototype.toObject = function(opt_includeInstance) {
@@ -79,8 +51,8 @@ proto.main.HardwareStats.prototype.toObject = function(opt_includeInstance) {
 
 /**
  * Static version of the {@see toObject} method.
- * @param {boolean|undefined} includeInstance Deprecated. Whether to include
- *     the JSPB instance for transitional soy proto support:
+ * @param {boolean|undefined} includeInstance Whether to include the JSPB
+ *     instance for transitional soy proto support:
  *     http://goto/soy-param-migration
  * @param {!proto.main.HardwareStats} msg The msg instance to transform.
  * @return {!Object}
@@ -201,12 +173,9 @@ proto.main.HardwareStats.prototype.getCpu = function() {
 };
 
 
-/**
- * @param {number} value
- * @return {!proto.main.HardwareStats} returns this
- */
+/** @param {number} value */
 proto.main.HardwareStats.prototype.setCpu = function(value) {
-  return jspb.Message.setProto3IntField(this, 1, value);
+  jspb.Message.setField(this, 1, value);
 };
 
 
@@ -219,12 +188,9 @@ proto.main.HardwareStats.prototype.getMemoryFree = function() {
 };
 
 
-/**
- * @param {number} value
- * @return {!proto.main.HardwareStats} returns this
- */
+/** @param {number} value */
 proto.main.HardwareStats.prototype.setMemoryFree = function(value) {
-  return jspb.Message.setProto3IntField(this, 2, value);
+  jspb.Message.setField(this, 2, value);
 };
 
 
@@ -237,29 +203,41 @@ proto.main.HardwareStats.prototype.getMemoryUsed = function() {
 };
 
 
-/**
- * @param {number} value
- * @return {!proto.main.HardwareStats} returns this
- */
+/** @param {number} value */
 proto.main.HardwareStats.prototype.setMemoryUsed = function(value) {
-  return jspb.Message.setProto3IntField(this, 3, value);
+  jspb.Message.setField(this, 3, value);
 };
 
 
 
+/**
+ * Generated by JsPbCodeGenerator.
+ * @param {Array=} opt_data Optional initial data array, typically from a
+ * server response, or constructed directly in Javascript. The array is used
+ * in place and becomes part of the constructed object. It is not cloned.
+ * If no data is provided, the constructed object will be empty, but still
+ * valid.
+ * @extends {jspb.Message}
+ * @constructor
+ */
+proto.main.EmptyRequest = function(opt_data) {
+  jspb.Message.initialize(this, opt_data, 0, -1, null, null);
+};
+goog.inherits(proto.main.EmptyRequest, jspb.Message);
+if (goog.DEBUG && !COMPILED) {
+  proto.main.EmptyRequest.displayName = 'proto.main.EmptyRequest';
+}
 
 
 if (jspb.Message.GENERATE_TO_OBJECT) {
 /**
- * Creates an object representation of this proto.
+ * Creates an object representation of this proto suitable for use in Soy templates.
  * Field names that are reserved in JavaScript and will be renamed to pb_name.
- * Optional fields that are not set will be set to undefined.
  * To access a reserved field use, foo.pb_<name>, eg, foo.pb_default.
  * For the list of reserved names please see:
- *     net/proto2/compiler/js/internal/generator.cc#kKeyword.
- * @param {boolean=} opt_includeInstance Deprecated. whether to include the
- *     JSPB instance for transitional soy proto support:
- *     http://goto/soy-param-migration
+ *     com.google.apps.jspb.JsClassTemplate.JS_RESERVED_WORDS.
+ * @param {boolean=} opt_includeInstance Whether to include the JSPB instance
+ *     for transitional soy proto support: http://goto/soy-param-migration
  * @return {!Object}
  */
 proto.main.EmptyRequest.prototype.toObject = function(opt_includeInstance) {
@@ -269,8 +247,8 @@ proto.main.EmptyRequest.prototype.toObject = function(opt_includeInstance) {
 
 /**
  * Static version of the {@see toObject} method.
- * @param {boolean|undefined} includeInstance Deprecated. Whether to include
- *     the JSPB instance for transitional soy proto support:
+ * @param {boolean|undefined} includeInstance Whether to include the JSPB
+ *     instance for transitional soy proto support:
  *     http://goto/soy-param-migration
  * @param {!proto.main.EmptyRequest} msg The msg instance to transform.
  * @return {!Object}
